# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elgiganten Repair API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/statuser": {
            "get": {
                "tags": [
                    "Sager - Statuser"
                ],
                "summary": "Liste af Statuser",
                "description": "",
                "operationId": "statusesList",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "navn": {
                                            "type": "string",
                                            "example": "Bedste Status Navn"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Status Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    },
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Status Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Statuser"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sager - Statuser"
                ],
                "summary": "Opret en ny Status",
                "description": "",
                "operationId": "statusesCreate",
                "parameters": [
                    {
                        "name": "navn",
                        "in": "query",
                        "description": "Status navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Status oprettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/statuser/{id}": {
            "get": {
                "tags": [
                    "Sager - Statuser"
                ],
                "summary": "Find en specifik Status",
                "description": "",
                "operationId": "statusesSpecific",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "navn": {
                                            "type": "string",
                                            "example": "Bedste Status Navn"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Status Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sager - Statuser"
                ],
                "summary": "Opdater en Status",
                "description": "",
                "operationId": "statusesUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "navn",
                        "in": "query",
                        "description": "Status navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status Opdateret"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sager - Statuser"
                ],
                "summary": "Slet en Status",
                "description": "",
                "operationId": "statusesDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Status Slettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}